+++ Chess.Test/Bitboard/BitboardTests.cs
@@ -24,16 +24,26 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-namespace Chess.Tests
+namespace Chess.Test.Bitboard
 {
-    public interface IPieceAttacks
+    using Rudz.Chess.Enums;
+    using Rudz.Chess.Types;
+    using Xunit;
+
+    public sealed class BitboardTests
     {
-        void AlphaPattern();
+        [Fact]
+        public void MakeBitBoardTest()
+        {
+            // a few squares
+            BitBoard b1 = BitBoards.MakeBitboard(ESquare.a1, ESquare.b1, ESquare.a2, ESquare.b2);
+            BitBoard b2 = ESquare.a1.BitBoardSquare() | ESquare.b1.BitBoardSquare() | ESquare.a2.BitBoardSquare() | ESquare.b2.BitBoardSquare();
+            Assert.Equal(b1, b2);
 
-        void BetaPattern();
-
-        void GammaPattern();
-
-        void DeltaPattern();
+            // a single square (not needed, but still has to work in case of list of squares etc)
+            BitBoard b3 = BitBoards.MakeBitboard(ESquare.h3);
+            BitBoard b4 = ESquare.h3.BitBoardSquare();
+            Assert.Equal(b3, b4);
+        }
     }
 }
