+++ Chess.Test/Pieces/PieceAttacksBishopTests.cs
@@ -24,25 +24,25 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-using Rudz.Chess.Enums;
-using Rudz.Chess.Types;
-
-namespace Chess.Tests
+namespace Chess.Test.Pieces
 {
+    using Rudz.Chess;
+    using Rudz.Chess.Types;
     using Xunit;
 
-    public class PieceAttacksRookTests : PieceAttacksSliders
+    /// <inheritdoc />
+    public sealed class PieceAttacksBishopTests : PieceAttacksSliders
     {
         [Fact]
         public override void AlphaPattern()
         {
             const int index = (int)EBands.Alpha;
-            const int sliderIndex = 1;
+            const int sliderIndex = 0;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(RookExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(BishopExpected[index], attacks.Count);
             }
         }
 
@@ -50,12 +50,12 @@ namespace Chess.Tests
         public override void BetaPattern()
         {
             const int index = (int)EBands.Beta;
-            const int sliderIndex = 1;
+            const int sliderIndex = 0;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(RookExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(BishopExpected[index], attacks.Count);
             }
         }
 
@@ -63,12 +63,12 @@ namespace Chess.Tests
         public override void GammaPattern()
         {
             const int index = (int)EBands.Gamma;
-            const int sliderIndex = 1;
+            const int sliderIndex = 0;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(RookExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(BishopExpected[index], attacks.Count);
             }
         }
 
@@ -76,47 +76,44 @@ namespace Chess.Tests
         public override void DeltaPattern()
         {
             const int index = (int)EBands.Delta;
-            const int sliderIndex = 1;
+            const int sliderIndex = 0;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(RookExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(BishopExpected[index], attacks.Count);
             }
         }
 
         /// <summary>
-        /// Testing results of blocked rook attacks, they should always return 7 on the sides, and 14 in the corner
+        /// Testing results of blocked bishop attacks, they should always return 2 on the sides, and 1 in the corner
         /// </summary>
         [Fact]
-        public void RookBorderBlocked()
+        public void BishopBorderBlocked()
         {
-            /*
-             * Test purpose : Testing blocked bishop attacks
-             */
             BitBoard border = 0xff818181818181ff;
             BitBoard borderInner = 0x7e424242427e00;
             BitBoard corners = 0x8100000000000081;
 
-            const int expectedCorner = 14;
-            const int expectedSide = 8; // 7 to each side and 1 blocked
+            const int expectedCorner = 1; // just a single attack square no matter what
+            const int expectedSide = 2;
 
             /*
-             * borderInner (X = set bit) :
-             *
-             * 0 0 0 0 0 0 0 0
-             * 0 X X X X X X 0
-             * 0 X 0 0 0 0 X 0
-             * 0 X 0 0 0 0 X 0
-             * 0 X 0 0 0 0 X 0
-             * 0 X 0 0 0 0 X 0
-             * 0 X X X X X X 0
-             * 0 0 0 0 0 0 0 0
-             *
-             */
-            foreach (Square square in border)
+                         * borderInner (X = set bit) :
+                         *
+                         * 0 0 0 0 0 0 0 0
+                         * 0 X X X X X X 0
+                         * 0 X 0 0 0 0 X 0
+                         * 0 X 0 0 0 0 X 0
+                         * 0 X 0 0 0 0 X 0
+                         * 0 X 0 0 0 0 X 0
+                         * 0 X X X X X X 0
+                         * 0 0 0 0 0 0 0 0
+                         *
+                         */
+            foreach (var square in border)
             {
-                BitBoard attacks = square.GetAttacks(EPieceType.Rook, borderInner);
+                var attacks = square.BishopAttacks(borderInner);
                 Assert.False(attacks.Empty());
                 Assert.Equal(corners & square ? expectedCorner : expectedSide, attacks.Count);
             }
