+++ Chess.Test/Pieces/PieceAttacksRookTests.cs
@@ -24,27 +24,24 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-
-using Rudz.Chess;
-using Rudz.Chess.Types;
-
-namespace Chess.Tests
+namespace Chess.Test.Pieces
 {
+    using Rudz.Chess.Enums;
+    using Rudz.Chess.Types;
     using Xunit;
 
-    /// <inheritdoc />
-    public class PieceAttacksBishopTests : PieceAttacksSliders
+    public sealed class PieceAttacksRookTests : PieceAttacksSliders
     {
         [Fact]
         public override void AlphaPattern()
         {
             const int index = (int)EBands.Alpha;
-            const int sliderIndex = 0;
+            const int sliderIndex = 1;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(BishopExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(RookExpected[index], attacks.Count);
             }
         }
 
@@ -52,12 +49,12 @@ namespace Chess.Tests
         public override void BetaPattern()
         {
             const int index = (int)EBands.Beta;
-            const int sliderIndex = 0;
+            const int sliderIndex = 1;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(BishopExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(RookExpected[index], attacks.Count);
             }
         }
 
@@ -65,12 +62,12 @@ namespace Chess.Tests
         public override void GammaPattern()
         {
             const int index = (int)EBands.Gamma;
-            const int sliderIndex = 0;
+            const int sliderIndex = 1;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(BishopExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(RookExpected[index], attacks.Count);
             }
         }
 
@@ -78,44 +75,47 @@ namespace Chess.Tests
         public override void DeltaPattern()
         {
             const int index = (int)EBands.Delta;
-            const int sliderIndex = 0;
+            const int sliderIndex = 1;
 
-            foreach (Square pieceLocation in Bands[index])
+            foreach (var pieceLocation in Bands[index])
             {
-                BitBoard attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
-                Assert.Equal(BishopExpected[index], attacks.Count);
+                var attacks = SlideAttacks[sliderIndex](pieceLocation, EmptyBoard);
+                Assert.Equal(RookExpected[index], attacks.Count);
             }
         }
 
         /// <summary>
-        /// Testing results of blocked bishop attacks, they should always return 2 on the sides, and 1 in the corner
+        /// Testing results of blocked rook attacks, they should always return 7 on the sides, and 14 in the corner
         /// </summary>
         [Fact]
-        public void BishopBorderBlocked()
+        public void RookBorderBlocked()
         {
+            /*
+             * Test purpose : Testing blocked bishop attacks
+             */
             BitBoard border = 0xff818181818181ff;
             BitBoard borderInner = 0x7e424242427e00;
             BitBoard corners = 0x8100000000000081;
 
-            const int expectedCorner = 1; // just a single attack square no matter what
-            const int expectedSide = 2;
+            const int expectedCorner = 14;
+            const int expectedSide = 8; // 7 to each side and 1 blocked
 
             /*
-                         * borderInner (X = set bit) :
-                         *
-                         * 0 0 0 0 0 0 0 0
-                         * 0 X X X X X X 0
-                         * 0 X 0 0 0 0 X 0
-                         * 0 X 0 0 0 0 X 0
-                         * 0 X 0 0 0 0 X 0
-                         * 0 X 0 0 0 0 X 0
-                         * 0 X X X X X X 0
-                         * 0 0 0 0 0 0 0 0
-                         *
-                         */
-            foreach (Square square in border)
+             * borderInner (X = set bit) :
+             *
+             * 0 0 0 0 0 0 0 0
+             * 0 X X X X X X 0
+             * 0 X 0 0 0 0 X 0
+             * 0 X 0 0 0 0 X 0
+             * 0 X 0 0 0 0 X 0
+             * 0 X 0 0 0 0 X 0
+             * 0 X X X X X X 0
+             * 0 0 0 0 0 0 0 0
+             *
+             */
+            foreach (var square in border)
             {
-                BitBoard attacks = square.BishopAttacks(borderInner);
+                var attacks = square.GetAttacks(EPieceType.Rook, borderInner);
                 Assert.False(attacks.Empty());
                 Assert.Equal(corners & square ? expectedCorner : expectedSide, attacks.Count);
             }
